Improved support for file names with long numbers
If you were in a folder with a file that had a really long number in
it's name (20 digits or more); vimage would crash when you tried to
navigate to another image. I fixed that by making the code that compares
names trim down numbers to 19 characters if they are too long.