Fit to Monitor AUTO option
Fit to Monitor now has an AUTO option which is now used by the middle
mouse button by default. Auto means the image will fit to monitor height
if it's on a horizontal monitor and fit to monitor width if it's on a
vertical one.

This is mainly an addition for me, since my monitor can rotate, which I
like to do when I view comic/manga pages.