Added ExitAllInstances function + taskbar hiding cmds
Someone wanted to be able to hide vimage from the taskbar. I messed around with an option to put vimage in the tray but it just got too messy when trying to get it to play nice with multiple instances. I don't really see why anyone would want to put vimage in the tray anyway so I scrapped that. While I was at it I thought I might as well put in some simple commands to hide/show vimage from the taskbar for people who want them. To use them just create a new Custom Action and put the command you want in there; then you can bind it to a key or put it on the context menu.