Added support for viewing Clipboard images
Opening vimage directly with an image in the clipboard will now display that image. Useful if just want to quickly take a screenshot (or snip with WIN+SHIFT+S) and view it.

eg: Take a screenshot of a game map/guide, open vimage and lock the image in place over the game.