Fixed possible crash + some refactoring
Someone had an ArgumentOutOfRangeException crash when vimage tried to get the folder contents. Not sure exactly how, though it seems like the backslash in the file path was gone (it said the filename was "G:VvNhMb0.jpg"). Possibly a problem with how windows deals with backslashes in file paths in different lanaguages?
Simply replaced the Substring/LastIndexOf with Path.GetDirectoryName.

Added a try-catch to Process.Start for Custom Actions to prevent the crash reporter from opening if a user messed something up.

Also replaced GetExtension with built-in .NET Path.GetExtension.